Introduce a configration that points a public endpoint for the search host.
The search-host is accessed both by the admin-backend, and directly by the
administrative-users browser. For situations where the public url for
the search-host is not accessible to the backend (eg. due to firewalls) we need
two seperate configurations for the public url (for the users browser) and the
internal url (for backend interaction).